Refactor: Project cleanup and enhancements
- Refactored test files: moved mocks, ensured _test package suffix.
- Enhanced MongoDB integration tests: ensured MONGO_TEST_URI check.
- Updated README.md: improved Quick Example, added TODO section.
- Project cleanup: relocated files, updated imports, added security headers, improved logging/error handling.

Refactor: Project cleanup and enhancements

- Refactored test files: moved mocks, ensured _test package suffix.
- Enhanced MongoDB integration tests: ensured MONGO_TEST_URI check.
- Updated README.md: improved Quick Example, added TODO section.
- Project cleanup: relocated files, updated imports, added security headers, improved logging/error handling.

Feat: Initial project refactor, app structure, and config setup

This commit includes a series of refactoring and setup steps:

1.  **Test File Refactoring**:
    *   Moved all mock files to a root `mocks/` directory.
    *   Regenerated mocks using `mockgen`.
    *   Ensured all test files use `_test` package suffixes (e.g., `mongodb_test`)
        and updated import paths accordingly.

2.  **MongoDB Integration Test Enhancements**:
    *   Verified that all MongoDB integration tests check for the `MONGO_TEST_URI`
        environment variable and skip if not set.

3.  **README.md Updates**:
    *   Rewrote the "Quick Example" section to be more accurate, focusing on library usage initially.
    *   Added a "TODO" section for future enhancements.
    *   Later, updated "Getting Started" to include instructions for running the new
        server application and `ssoctl` CLI, including Viper configuration details.

4.  **Project Cleanup and Initial Production Readiness**:
    *   Relocated several root-level Go files (`jwks_service.go`, `keys.go`, etc.)
        into more appropriate packages (`services/`, `internal/crypto/`).
    *   Updated import paths across the project to reflect these moves.
    *   Corrected package declarations in moved files (e.g., ensuring `package services` for files in the `services` directory).
    *   Resolved numerous import cycles and undefined type/variable errors by:
        *   Centralizing domain types (Token, TokenInfo, AuthCode, DeviceCode, etc.) into the `domain` package.
        *   Defining repository interfaces (OAuthRepository, TokenRepository, etc.) in `domain/repositories.go`.
        *   Moving context utility functions to `domain/context.go`.
        *   Consolidating error variable definitions into `errors/oauth_errors.go`.
        *   Updating services and repository implementations to use these centralized domain types, interfaces, and errors.
    *   Added a basic security headers middleware for the Gin API.
    *   Made minor improvements to logging by removing potentially sensitive information in a few specific cases.

5.  **Application Restructuring**:
    *   Created an `apps/` directory.
    *   Moved the main SSO server logic to `apps/ssso/ssso.go`.
    *   Moved the `ssoctl` CLI application from `cmd/ssoctl/` to `apps/ssoctl/`,
        renaming its entry point to `sssoctl.go` and restructuring its internal
        directories.
    *   Updated all relevant import paths for `ssoctl`.
    *   The old `cmd/` directory at the project root has been cleaned up.

6.  **Configuration with Viper (SSO Server)**:
    *   Integrated Viper into `apps/ssso/ssso.go` for configuration management.
    *   Defined a `Config` struct and `LoadConfig` function to read from `sso_config.yaml` or environment variables (prefixed with `SSSO_`).
    *   Modified `internal/server/connectrpc_server.go` to accept a `ServerConfig` struct, allowing it to use Viper-loaded configurations for HTTP address, MongoDB URI/DBName, and Issuer URL.

7.  **CLI with Cobra (`ssoctl`)**:
    *   Verified that `ssoctl` (now in `apps/ssoctl/`) already uses Cobra.
    *   Confirmed its existing Viper integration for configuration (`~/.ssoctl/config.yaml`) remains functional after the move.

8.  **Installable Applications**:
    *   Ensured both `apps/ssso/ssso.go` and `apps/ssoctl/sssoctl.go` are installable via `go install` after resolving build errors.

**Current State & Next Steps (from interrupted plan):**

The immediate next step was to continue the "Project cleanup and production readiness (re-evaluate)" phase, specifically by focusing on robust RSA key management for token signing. This involved refactoring `services.TokenSigner` and `services.JWKSService` to load/generate RSA keys from a path specified in the configuration (`SSSO_SIGNING_KEY_PATH`) and ensuring `TokenService` uses these RSA keys for signing tokens. Further steps included correctly handling client secret hashing comparisons and potentially refactoring `OAuthService` to reduce redundancy with `AuthService`.

I was in the process of requesting user feedback on this key management plan when the session ended.

feat: Add Docker, GitHub Actions CI, and Helm chart

- Add Dockerfile for building the ssso application.
- Add GitHub Actions workflow for build, test, coverage, and Docker image build on PRs and main branch.
- Add GitHub Actions workflow for publishing Docker image to GHCR on push to v1 branch.
- Create Helm chart for Kubernetes deployment, including:
  - Deployment, Service, HPA, Ingress
  - Configurable values for ingress (Nginx/Traefik), TLS (CertManager), MongoDB, and application settings.
  - PersistentVolumeClaim for optional storage.
  - Guidance for MongoDB integration (recommends subchart).
- Add /healthz and /readyz HTTP endpoints to the Go application for Kubernetes probes.

feat: Add Docker, GitHub Actions CI, and Helm chart

- Add Dockerfile for building the ssso application.
- Add GitHub Actions workflow for build, test, coverage, and Docker image build on PRs and main branch.
- Add GitHub Actions workflow for publishing Docker image to GHCR on push to v1 branch.
- Create Helm chart for Kubernetes deployment, including:
  - Deployment, Service, HPA, Ingress
  - Configurable values for ingress (Nginx/Traefik), TLS (CertManager), MongoDB, and application settings.
  - PersistentVolumeClaim for optional storage.
  - Guidance for MongoDB integration (recommends subchart).
- Add /healthz and /readyz HTTP endpoints to the Go application for Kubernetes probes.

docs: Add ENHANCEMENTS_TODO.md for future improvements

docs: Add specification for Distributed Token Store service

Defines the architecture, gRPC API, and implementation considerations for a BBoltDB-based gRPC service to act as a high-performance store for sessions, flows, and tokens as an alternative to Redis or direct DB calls.

docs: Add DISTRIBUTED_TOKEN_STORE_SPEC.md

This file was previously created but not found in the repository state. Re-adding it.

docs: Add ENHANCEMENTS_TODO.md

This file outlines future improvement ideas for the SSSO backend project. Re-adding as it was not found in the previous repository state.

Removed testify based shitty mocks
