--- apps/ssoctl/sssoctl.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/pilab-dev/shadow-sso/apps/ssoctl/cmd" // Updated import path for Execute
-)
-
-func main() {
-	cmd.Execute()
-}
