+++ apps/sssoctl/cmd/config/config.go
