+++ apps/sssoctl/cmd/context.go
@@ -4,7 +4,7 @@ import (
 	"errors" // Added for setContextCmd
 	"fmt"
 
-	"github.com/pilab-dev/shadow-sso/apps/ssoctl/cmd/config"
+	"github.com/pilab-dev/shadow-sso/apps/sssoctl/cmd/config"
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v3" // For pretty printing contexts
 )
