+++ apps/sssoctl/cmd/idp.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 
 	"connectrpc.com/connect"
-	"github.com/pilab-dev/shadow-sso/apps/ssoctl/cmd/client"
-	"github.com/pilab-dev/shadow-sso/apps/ssoctl/cmd/config"
+	"github.com/pilab-dev/shadow-sso/apps/sssoctl/cmd/client"
+	"github.com/pilab-dev/shadow-sso/apps/sssoctl/cmd/config"
 	ssov1 "github.com/pilab-dev/shadow-sso/gen/proto/sso/v1"
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v3"
