+++ apps/sssoctl/sssoctl.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/pilab-dev/shadow-sso/apps/sssoctl/cmd" // Updated import path for Execute
+)
+
+func main() {
+	cmd.Execute()
+}
